RDB-64 feat: created EnhancedTableHead component
